requirements.txt
fastapi
uvicorn
langchain
openai
pydantic
python-dotenv
tiktoken
chromadb
playwright
pymupdf


langchain-community
